Rename backend public access filter for better clarity
