+++ symphony-ws/src/main/java/se/havochvatten/symphony/web/filter/PublicOrRestricted.java
@@ -1,4 +1,4 @@
-package se.havochvatten.symphony.web.publicfilter;
+package se.havochvatten.symphony.web.filter;
 
 import jakarta.ws.rs.NameBinding;
 import java.lang.annotation.Retention;
@@ -10,6 +10,6 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @NameBinding
 @Retention(RUNTIME)
 @Target({TYPE, METHOD})
-public @interface ConditionalPublic {
+public @interface PublicOrRestricted {
     String[] roles() default {};
 }
